refactored to create new spaceship part component. Brought back tyde menu home for future use. Addressed few other PR comments
