+++ templates/vue/src/components/tyde/TydeMenu.vue
@@ -31,7 +31,6 @@
     <div class="content">
       <tyde-cockpit
         v-if="activePage === 'home'"
-        :logs="logs"
         @close="$emit('continue')"
       />
       <tyde-menu-settings
